backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.15.0"
fastapi = "^0.68.1"
python-multipart = "^0.0.5"
email-validator = "^1.0.5" # TO REMOVE?
requests = "^2.23.0"
celery = "^4.4.2" # TO REMOVE
passlib = {extras = ["bcrypt"], version = "^1.7.2"} # TO REMOVE
tenacity = "^6.1.0"
pydantic = "^1.4"
emails = "^0.5.15" # TO REMOVE
raven = "^6.10.0" # TO REMOVE
gunicorn = "^20.0.4"
jinja2 = "^2.11.2"
pytest = "^5.4.1"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
orcid = "^1.0.3"
motor = "^2.5.1"
Authlib = "^0.15.4"
itsdangerous = "^2.0.1"
httpx = "0.18.1"
extruct = "^0.13.0"
rdflib = "^6.0.1"
PyLD = "^2.0.3"
google = "^3.0.0"
beautifulsoup4 = "^4.10.0"
strawberry-graphql = "^0.81.0"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

